refactor(cli): extract persistence lifecycle to _persist.py
Consolidates ~300 lines of duplicated DB session/branch creation,
hook registration, and teardown logic from agent.py and
orchestrate/_orchestration.py into a single canonical module.
Moves resolve_run_reason out of agent.py to break the
_orchestration.py → agent.py import cycle.
